Extract test of typeclasses.
